requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.14.1
amqp==5.3.1
anyio==4.8.0
asgiref==3.11.0
attrs==25.1.0
Authlib==1.4.0
autobahn==25.12.2
Automat==25.4.16
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-mgmt-core==1.3.2
azure-mgmt-rdbms==10.2.0b17
azure-mgmt-resource==23.1.1
azure-mgmt-subscription==3.0.0
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.14.2
bidict==0.23.1
billiard==4.2.4
blinker==1.9.0
boto3==1.40.56
botocore==1.40.56
Brotli==1.1.0
cachetools==5.5.1
cbor2==5.7.1
celery==5.6.0
certifi==2025.7.9
cffi==2.0.0
channels==4.3.2
channels-redis==4.3.0
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==44.0.0
decorator==5.1.1
defusedxml==0.7.1
Django==5.2.9
django-allauth==65.13.1
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.2
django-otp==1.6.3
django-redis==6.0.0
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.1
flower==2.0.1
frozenlist==1.8.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
humanize==4.12.0
hyperlink==21.0.0
idna==3.10
Incremental==24.11.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.6.1
ldap3==2.9
lxml==5.3.2
Mako==1.2.3
MarkupSafe==3.0.2
msgpack==1.1.2
multidict==6.1.0
oauthlib==3.2.2
packaging==24.2
passlib==1.7.4
phonenumbers==9.0.21
pillow==11.1.0
prometheus_client==0.23.1
protobuf==3.19.6
psutil==6.1.1
psycopg2-binary==2.9.11
PyJWT==2.8.0
pyOpenSSL==25.0.0
pyotp==2.9.0
python-crontab==3.3.0
python-dateutil==2.8.2
python-engineio==4.12.3
python-socketio==5.15.0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.0
RapidFuzz==3.11.0
razorpay==2.0.0
redis==7.1.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.6
requests==2.32.4
requests-oauthlib==1.3.1
rich==13.9.4
rpds-py==0.25.0
rsa==4.9
s3transfer==0.14.0
SecretStorage==3.3.3
sentry-sdk==2.21.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.4.2
stripe==14.1.0
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.12.2
tzdata==2025.3
ujson==5.11.0
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.22.0
zope.interface==8.1.1
zstandard==0.23.0
drf-yasg==1.21.7
setuptools>=68.0.0
daphne==4.2.1

